Refactor CreateOrUpdateLabel
Only expose state / finish callback.
